# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Cumora API
  version: 0.2.2
  description: |
    Cumora 单一契约源(ADR 0004 · 票 #48)。覆盖全部存活路由域:
    core(auth/me/companies/invitations)、computers/agents、hr(编外人事代理)、
    conversations/messages、email、projects、participants、boards、calendar、
    documents、push、uploads、projects、observability、devtools、admin、
    shipping、runtime(BYOA daemon 专用面)。

    ## WebSocket / SSE 事件面描述策略

    本规范描述 HTTP 面。长连接通道:

    - **WebSocket `/ws`** — 实时事件广播(message.new/delta、typing、participants/
      computers 状态、board/doc/calendar 变更等)。连接鉴权走 `/auth/ws-ticket`
      票据。事件载荷自 #221 起契约化:逐事件 schema 见同目录 `ws-events.json`
      (单一 JSON Schema 契约源,`npm run contract:gen` 再生 TS `src/ws-events.d.ts`
      与 Go `internal/events/ws.gen.go`,三端手写类型已退役)。
    - **SSE `/runtime/wake-stream`** — daemon 唤醒流(wake/steer 两类事件);载荷
      形状暂不入契约(#49 的 wire 级验收镜像测试承载——同一套测试跑 TS 与 Go,
      差异即回归)。

    WS 事件载荷不进本文件的 components:事件面走独立 JSON Schema 契约文件,
    避免把 $defs 级联混入 HTTP 生成管线(oapi-codegen / openapi-typescript 输出
    保持面干净);两文件通过 `openapi.yaml#/components/schemas/*` $ref 共享组件。

    ## 鉴权

    - `sessionBearer` — 用户会话(Authorization: Bearer <token> + X-Company-Id 租户头)
    - `agentRuntimeJWT` — BYOA daemon 的 per-agent runtime JWT(/runtime/* 全域)
    - `deviceToken` — daemon 设备凭证(/api/computers/pair 等)
    - 部分路由匿名(public/livez/og/auth 回调)
servers:
  - url: /
  - url: http://localhost:5181
...
security:
  - sessionBearer: []
paths:
...
  /api/projects:
    get:
      tags: [projects]
      operationId: listProjects
      summary: 项目列表
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [projects]
      operationId: createProject
      summary: 建项目
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                color: { type: string }
                folderPath:
                  type: string
                  description: 可选自填已有文件夹(缺省在受管目录自动建空盘,ADR 0008 §3)。
      responses:
        '201':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [id, name, description, color, status, folderPath, isDefault]
                properties:
                  id: { type: string }
                  name: { type: string }
                  description: { type: string }
                  color: { type: string, nullable: true }
                  status: { type: string }
                  folderPath: { type: string }
                  isDefault: { type: boolean }
  /api/projects/{id}:
    get:
      tags: [projects]
      operationId: getProject
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      summary: 项目详情(盘路径仅特权成员可见;#355 起原工作区详情面)
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectDetail' } } } }
    delete:
      tags: [projects]
      operationId: deleteProject
      summary: 删除项目(#354,ADR 0008 §6 生命周期唯一出口)
      description: 对话 SET NULL 保留、交付台账随卡片存活(FK SET NULL)、成员/关联清理;盘文件原地保留(不代删);默认项目拒删。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [ok, id, folderKept]
                properties:
                  ok: { type: boolean }
                  id: { type: string }
                  folderKept: { type: string, nullable: true }
        '403': { description: 默认项目不可删 / 非管理员 }
        '404': { description: not found }
    put:
      tags: [projects]
      operationId: updateProject
      summary: 改项目
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                color: { type: string, nullable: true }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
...
  /api/projects/{id}/members:
    post:
      tags: [projects]
      operationId: addProjectMember
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      summary: 加显式成员
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [participantId]
              properties:
                participantId: { type: string }
      responses:
        '201': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/projects/{id}/members/{participantId}:
    delete:
      tags: [projects]
      operationId: removeProjectMember
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: participantId, in: path, required: true, schema: { type: string } }
      summary: 移除显式成员
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/projects/{id}/associations:
    post:
      tags: [projects]
      operationId: addProjectAssociation
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      summary: 关联(看板卡/文档;project-kind 已随概念合并退役 #354)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [kind, targetId]
              properties:
                kind: { type: string, enum: [board_card, document] }
                targetId: { type: string }
      responses:
        '201':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [ok, kind, targetId]
                properties:
                  ok: { type: boolean }
                  kind: { type: string, enum: [board_card, document] }
                  targetId: { type: string }
  /api/projects/{id}/associations/{kind}/{targetId}:
    delete:
      tags: [projects]
      operationId: removeProjectAssociation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        # 枚举校验与文案在 handler('kind must be one of board_card, document')
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: targetId, in: path, required: true, schema: { type: string } }
      summary: 解除关联
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/projects/{id}/files:
    get:
      tags: [projects]
      operationId: listProjectFiles
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        # 校验与文案在 handler('path required' 或作可选过滤),保持宽松声明
        - { name: path, in: query, schema: { type: string } }
      summary: 盘内列目录(范围制)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [path, entries]
                properties:
                  path: { type: string }
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectFileEntry' }
  /api/projects/{id}/file:
    get:
      tags: [projects]
      operationId: readProjectFile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        # 校验与文案在 handler('path required' 或作可选过滤),保持宽松声明
        - { name: path, in: query, schema: { type: string } }
      summary: 读文件
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [path, body, size, modifiedAt]
                properties:
                  path: { type: string }
                  body: { type: string }
                  size: { type: integer }
                  modifiedAt: { type: string, format: date-time }
                  mtimeNanos: { type: string, description: '#337 纳秒 mtime(十进制字符串,CAS expected 用)' }
    put:
      tags: [projects]
      operationId: writeProjectFile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        # 校验与文案在 handler('path required' 或作可选过滤),保持宽松声明
        - { name: path, in: query, schema: { type: string } }
        # #337 CAS:可选乐观并发(纳秒 mtime,十进制字符串 —— int64 超 JS
        # 安全整数)。失配 412 + 挑战者内容留 .conflict 副本。
        - { name: expectedMtimeNanos, in: query, schema: { type: string } }
      summary: 写文件(#337 起支持 CAS)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  path: { type: string }
                  mtimeNanos: { type: string, description: 写后 mtime(下一轮 CAS 的 expected 用) }
        '412':
          description: CAS 失配(盘上 mtime 与 expected 不符;挑战者内容已留 .conflict 副本)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  currentMtimeNanos: { type: string }
                  conflictPath: { type: string }
  /api/projects/{id}/upload:
    post:
      tags: [projects]
      operationId: uploadProjectFile
      summary: '#338 multipart 上传二进制(单文件 25MB 帽;复用防逃逸/保留路径/写前快照)'
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [path, file]
              properties:
                path: { type: string, description: 盘内相对路径 }
                file: { type: string, format: binary }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  path: { type: string }
                  size: { type: integer }
                  mtimeNanos: { type: string }
        '413':
          description: 超过 25MB 二进制帽
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /api/projects/{id}/raw:
    get:
      tags: [projects]
      operationId: readProjectFileRaw
      summary: '#338 原始字节读(图片预览/下载;Content-Type 按扩展名猜)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: 原始文件字节(图片扩展名回对应 image/*,其余 octet-stream)
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            image/png:
              schema: { type: string, format: binary }
            image/jpeg:
              schema: { type: string, format: binary }
            image/gif:
              schema: { type: string, format: binary }
            image/webp:
              schema: { type: string, format: binary }
            image/svg+xml:
              schema: { type: string, format: binary }
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      description: 用户会话令牌;多租户路由另需 X-Company-Id 头
    agentRuntimeJWT:
      type: http
      scheme: bearer
      description: daemon 经 /api/agents/:id/runtime-token 领取的 per-agent JWT(2h)
    deviceToken:
      type: http
      scheme: bearer
      description: daemon 配对所得设备凭证(computers.pair 发放)
  schemas:
...
    Project:
      type: object
      required: [id, name, description, color, status, createdAt, archivedAt, folderPath, isDefault, conversationCount, explicitMemberCount]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        color: { type: string, nullable: true }
        status: { type: string, enum: [active, archived] }
        createdAt: { type: string, format: date-time }
        archivedAt: { type: string, nullable: true, format: date-time }
        folderPath:
          type: string
          nullable: true
          description: 项目绑定的真实文件夹(ADR 0008 §3;存量无盘项目补盘前为 null)。
        isDefault: { type: boolean }
        conversationCount: { type: integer }
        explicitMemberCount: { type: integer, description: '显式成员数(推导成员不计;#355 并入原列表语义)' }
...
    ProjectDetail:
      type: object
      required: [id, name, isDefault, createdAt, members, associations]
      properties:
        id: { type: string }
        name: { type: string }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
        folderPath: { type: string }
        members:
          type: array
          items: { $ref: '#/components/schemas/ProjectMember' }
        associations:
          type: array
          items: { $ref: '#/components/schemas/ProjectAssociation' }
...